refactor(FSM): Separacion de responsabilidades en Monsters y Player
