+++ src/roguelike_game/ecs/systems/input/input_system.py
@@ -10,13 +10,10 @@ from roguelike_game.ecs.components.combat.attack_cooldown import AttackCooldown
 from roguelike_game.ecs.components.ai.wants_to_melee import WantsToMelee
 from roguelike_game.ecs.components.ai.wants_to_cast import WantsToCastSpell
 from roguelike_engine.utils.benchmark import benchmark
-from roguelike_game.ecs.fsm.states.aggro_state import AggroState
 from roguelike_game.ecs.fsm.states.idle_state import IdleState
 from roguelike_game.ecs.fsm.states.player.move_state import MoveState
 from roguelike_game.ecs.fsm.states.player.player_attack_state import PlayerAttackState
 from roguelike_game.ecs.fsm.states.player.player_spell_select_state import PlayerSpellSelectState
-from roguelike_game.ecs.fsm.states.player.player_spell_cast_state import PlayerSpellCastState
-from roguelike_game.ecs.components.core.player_tag import PlayerTagComponent
 from roguelike_game.ecs.systems.fsm.fsm_system import _EntityProxy
 
 class InputSystem:
